Fixed 100s of nasty CDB parser bugs.
git-svn-id: https://jscsi.svn.sourceforge.net/svnroot/jscsi/trunk@267 6391b2ca-d626-0410-90c9-8f6d1efc1dcd

+++ projects/scsi/src/org/jscsi/scsi/protocol/cdb/Write16.java
@@ -50,7 +50,7 @@ public class Write16 extends Write10
 
       // CAUTION: Signed longs represent unsigned longs
       setLogicalBlockAddress(in.readLong());
-      setTransferLength(in.readLong());
+      setTransferLength(in.readInt());
 
       setGroupNumber(in.readUnsignedByte() & 0x1F);
       super.setControl(in.readUnsignedByte());
@@ -75,7 +75,7 @@ public class Write16 extends Write10
 
          // CAUTION: Signed longs represent unsigned longs
          out.writeLong(getLogicalBlockAddress());
-         out.writeLong(getTransferLength());
+         out.writeInt((int) getTransferLength());
 
          out.writeByte(getGroupNumber() & 0x1F);
          out.writeByte(super.getControl());
